refactor(pipeline): split pipeline.py into package, fix mypy, add CLI tests
- Split pipeline.py (1146 lines) into audiosmith/pipeline/ package:
  - core.py: DubbingPipeline orchestrator (674 lines)
  - tts_synthesis.py: TTSSynthesisMixin with 13 TTS methods (429 lines)
  - helpers.py: 6 pure utility functions (107 lines)
- Fix all 97 mypy errors (0 remaining), remove || true from CI
- Add 10 E2E CLI integration tests (transcribe, TTS, dub commands)
- Add CHANGELOG.md (v0.5.0 and v0.4.0 releases)
- Omit hardware-dependent modules from coverage measurement
- 1091 tests passing, ruff clean, bandit 0 high/medium, mypy 0 errors

+++ audiosmith/mixer.py
@@ -99,7 +99,7 @@ class AudioMixer:
             ).T
 
         if len(bg_audio) >= total_samples:
-            return bg_audio[:total_samples]
+            return bg_audio[:total_samples]  # type: ignore[no-any-return]
 
         buffer = np.zeros((total_samples, 2), dtype=np.float32)
         buffer[:len(bg_audio)] = bg_audio
@@ -196,7 +196,7 @@ class AudioMixer:
         if peak > 0.01:
             buffer = buffer / peak * 0.95
 
-        return buffer
+        return buffer.astype(np.float32)
 
     def render_to_file(
         self,
